initial attempt at refactoring nn stuff and model loading
