+++ madmom/ml/__init__.py
@@ -7,4 +7,4 @@ Machine learning package.
 from __future__ import absolute_import, division, print_function
 
 # import the submodules
-from . import rnn, hmm, gmm
+from . import hmm, gmm, nn
