+++ madmom/ml/nn/__init__.py
@@ -1,0 +1,2 @@
+from nets import RecurrentNeuralNetwork, RNNProcessor
+from . import layers, transfer_fns
