+++ madmom/ml/nn/layers.pxd
@@ -1,10 +1,10 @@
 # encoding: utf-8
 """
-This file contains the speed critical parts of recurrent neural network (RNN)
-related functionality.
+This file contains the speed critical parts of neural network related
+functionality.
 
 Note: right now, this file is just an empty augmenting file. However, it
-increases performance when cython is used to compile the normal rnn.py file.
+increases performance when cython is used to compile the normal layers.py file.
 
 """
