+++ setup.py
@@ -17,8 +17,15 @@ import numpy as np
 version = '0.14.dev'
 
 # define which extensions need to be compiled
-extensions = [Extension('madmom.ml.rnn',
-                        ['madmom/ml/rnn.py', 'madmom/ml/rnn.pxd'],
+extensions = [Extension('madmom.ml.nn',
+                        ['madmom/ml/nets.py', 'madmom/ml/nets.pxd'],
+                        include_dirs=[np.get_include()]),
+              Extension('madmom.ml.nn',
+                        ['madmom/ml/layers.py', 'madmom/ml/layers.pxd'],
+                        include_dirs=[np.get_include()]),
+              Extension('madmom.ml.nn',
+                        ['madmom/ml/transfer_fns.py',
+                         'madmom/ml/transfer_fns.pxd'],
                         include_dirs=[np.get_include()]),
               Extension('madmom.audio.comb_filters',
                         ['madmom/audio/comb_filters.pyx'],
