+++ tests/__init__.py
@@ -16,3 +16,4 @@ AUDIO_PATH = DATA_PATH + 'audio/'
 ACTIVATIONS_PATH = DATA_PATH + 'activations/'
 ANNOTATIONS_PATH = DATA_PATH + 'annotations/'
 DETECTIONS_PATH = DATA_PATH + 'detections/'
+MODELS_PATH = DATA_PATH + 'models/'
